% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generatenestedsubsetsC}
\alias{generatenestedsubsetsC}
\alias{generate_nested_subsetsC}
\title{Generate nested subsets (Rcpp)}
\usage{
generate_nested_subsetsC(set)
}
\arguments{
\item{set}{An integer vector.}
}
\value{
A list of nested subsets.
}
\description{
Generate nested subsets \{1\}, \{1,2\}, ..., \{1,...,n\} using C++.
}
