% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generatesubsetsC}
\alias{generatesubsetsC}
\alias{generate_subsetsC}
\title{Generate all non-empty subsets (Rcpp)}
\usage{
generate_subsetsC(set)
}
\arguments{
\item{set}{An integer vector.}
}
\value{
A list containing all non-empty subsets.
}
\description{
Generate all non-empty subsets of an input vector using C++.
}
