\name{SA24204150-package}
\alias{SA24204150-package}
\docType{package}
\title{This package provides several functions for survival analysis and model averaging.}
\description{R and Rcpp versions of functions for survival analysis and model averaging }
\author{
Mengyao Li
Maintainer: Mengyao Li <lmmy@mail.ustc.edu.cn>
}
\keyword{ package }